test(worker): move the SDK divergence pins to a dispatch-only suite
